Refactor : 디렉토리명 변경(socket, io)
상세:
- 현재 사용하는 socket 관련 디렉토리인 io는 socket으로 변경
- 기존의 socket 디렉토리는 old-socket으로 변경
