--- server/io/index.js
@@ -1,3 +1,0 @@
-const io = require('./init');
-
-module.exports = io;
