+++ server/socket/handlers/index.js
@@ -1,0 +1,3 @@
+const matchHandler = require('./matchHandler');
+
+module.exports = { matchHandler };
